WFLY-6827 Bump management API version to incorporate changes due to WFLY-6782
