+++ clustering/jgroups/extension/src/main/java/org/jboss/as/clustering/jgroups/subsystem/ChannelResourceDefinition.java
@@ -108,6 +108,8 @@ public class ChannelResourceDefinition extends ChildResourceDefinition {
                     .addRejectCheck(RejectAttributeChecker.DEFINED, Attribute.CLUSTER.getDefinition())
                     ;
         }
+
+        ForkProtocolResourceDefinition.buildTransformation(version, builder);
     }
 
     final boolean allowRuntimeOnlyRegistration;
